chore: disable issues and direct ppl to web-components repo [skip ci]
All component-related issues (for any version of the component) should be opened either in the `web-components` monorepo (if it's reproducible with TypeScript and HTML, without Vaadin Flow / Java), or in the `flow-components` monorepo (if it's about the Java API).

This PR disables new issues in this repo. Any attempt to create a new issue here would direct users either to the `web-components` or to the `flow-components` monorepo

Related to: https://github.com/vaadin/components-team-tasks/issues/584